// api/generated/model/completeStepRequest.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LgymApi.Api
 * OpenAPI spec version: 1.0
 */
import type { CompleteStepRequestStep } from './completeStepRequestStep';
import type { CompleteStepRequestTutorialType } from './completeStepRequestTutorialType';

export interface CompleteStepRequest {
  tutorialType?: CompleteStepRequestTutorialType;
  step?: CompleteStepRequestStep;
}