// api/generated/model/completeStepRequestTutorialType.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LgymApi.Api
 * OpenAPI spec version: 1.0
 */

export type CompleteStepRequestTutorialType = typeof CompleteStepRequestTutorialType[keyof typeof CompleteStepRequestTutorialType];


export const CompleteStepRequestTutorialType = {
  Unknown: 'Unknown',
  OnboardingDemo: 'OnboardingDemo',
} as const;