// api/generated/model/completeTutorialRequest.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LgymApi.Api
 * OpenAPI spec version: 1.0
 */
import type { CompleteTutorialRequestTutorialType } from './completeTutorialRequestTutorialType';

export interface CompleteTutorialRequest {
  tutorialType?: CompleteTutorialRequestTutorialType;
}