// api/generated/model/tutorialProgressDto.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LgymApi.Api
 * OpenAPI spec version: 1.0
 */
import type { TutorialProgressDtoCompletedStepsItem } from './tutorialProgressDtoCompletedStepsItem';
import type { TutorialProgressDtoRemainingStepsItem } from './tutorialProgressDtoRemainingStepsItem';
import type { TutorialProgressDtoTutorialType } from './tutorialProgressDtoTutorialType';

export interface TutorialProgressDto {
  /** @nullable */
  _id?: string | null;
  tutorialType?: TutorialProgressDtoTutorialType;
  /** @nullable */
  tutorialName?: string | null;
  /** @nullable */
  tutorialDescription?: string | null;
  isCompleted?: boolean;
  /** @nullable */
  completedAt?: string | null;
  /** @nullable */
  completedSteps?: TutorialProgressDtoCompletedStepsItem[] | null;
  /** @nullable */
  remainingSteps?: TutorialProgressDtoRemainingStepsItem[] | null;
  totalSteps?: number;
  completedStepsCount?: number;
}