// api/generated/model/tutorialProgressDtoCompletedStepsItem.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LgymApi.Api
 * OpenAPI spec version: 1.0
 */

export type TutorialProgressDtoCompletedStepsItem = typeof TutorialProgressDtoCompletedStepsItem[keyof typeof TutorialProgressDtoCompletedStepsItem];


export const TutorialProgressDtoCompletedStepsItem = {
  Unknown: 'Unknown',
  CreateArea: 'CreateArea',
  CreateGym: 'CreateGym',
  CreatePlan: 'CreatePlan',
  CreatePlanDay: 'CreatePlanDay',
  CreateTraining: 'CreateTraining',
  LastTreningResult: 'LastTreningResult',
} as const;