// api/generated/model/tutorialProgressDtoRemainingStepsItem.ts
/**
 * Generated by orval v8.2.0 🍺
 * Do not edit manually.
 * LgymApi.Api
 * OpenAPI spec version: 1.0
 */

export type TutorialProgressDtoRemainingStepsItem = typeof TutorialProgressDtoRemainingStepsItem[keyof typeof TutorialProgressDtoRemainingStepsItem];


export const TutorialProgressDtoRemainingStepsItem = {
  Unknown: 'Unknown',
  CreateArea: 'CreateArea',
  CreateGym: 'CreateGym',
  CreatePlan: 'CreatePlan',
  CreatePlanDay: 'CreatePlanDay',
  CreateTraining: 'CreateTraining',
  LastTreningResult: 'LastTreningResult',
} as const;